Update README.md
Black Mirror is a British anthology television series created by Charlie Brooker.
Individual episodes explore a diversity of genres, but most are set in near-future dystopias with science fiction technology—a type of speculative fiction. 
The series is based on The Twilight Zone and uses technology to comment on contemporary social issues.